Correct uses of the word "dependence" (#282)
The word "dependency" used to describe the relation between multiple
memory accesses are replaced by "dependence". Although both words can be
used to describe "the state of being dependent", the word "dependence"
is preferred in compiler literatures. The word "dependency" is used most
for "the thing that is dependent upon" in the context of software.