Fix version analysis for AD
We need to analyze versions of each variables to save it in tapes or to
recompute it. To save space and to do less recomputation, a version is
identified only when its value is required at the backward pass.

Previously, we only considered `x` for use in the backward pass of
`y = f(x)`. However, we overlooked that `y` may also be needed to compute
the gradient of `y = f(x)`. This oversight caused issues with our tape or
recomputation. I am fixing this issue in this PR.

To determine which `y = f(x)` statements use `y` in their gradient
calculation, we need to check each statement’s gradient in the first place.
Then we perform versioning and generate the gradient back again that uses
the taped or recomputed value. This process is implemented in this PR
through a lazy pass in `autograd/derivative.h`, which is upgraded from
the old `GradExpr`. `ReplaceBySaved` is also moved to a separated
`autograd/replcae_by_saved.h` file, to be shared by both
`autograd/grad.h` and `autograd/derivative.h`.

I have added a new test to verify the versioning of `y`. Additionally, I
have included another test to confirm that our lazy gradient does not
interfere with our error reporting.